+++ lib/src/mt_main.h
@@ -545,7 +545,7 @@ struct mt_admin {
 };
 
 struct mt_kport_info {
-  /* dpdk port name for kernel port */
+  /* dpdk port name for kernel port(MTL_PMD_DPDK_AF_XDP) */
   char port[MTL_PORT_MAX][MTL_PORT_MAX_LEN];
 };
 
@@ -843,6 +843,14 @@ enum mtl_port mt_port_by_id(struct mtl_main_impl* impl, uint16_t port_id);
 
 static inline uint8_t* mt_sip_addr(struct mtl_main_impl* impl, enum mtl_port port) {
   return mt_get_user_params(impl)->sip_addr[port];
+}
+
+static inline uint8_t* mt_sip_netmask(struct mtl_main_impl* impl, enum mtl_port port) {
+  return mt_get_user_params(impl)->netmask[port];
+}
+
+static inline uint8_t* mt_sip_gatway(struct mtl_main_impl* impl, enum mtl_port port) {
+  return mt_get_user_params(impl)->gateway[port];
 }
 
 static inline enum mtl_pmd_type mt_pmd_type(struct mtl_main_impl* impl,