+++ lib/src/mt_socket.h
@@ -7,7 +7,10 @@
 
 #include "mt_main.h"
 
-int mt_socket_get_if_ip(char* if_name, uint8_t ip[MTL_IP_ADDR_LEN]);
+int mt_socket_get_if_ip(char* if_name, uint8_t ip[MTL_IP_ADDR_LEN],
+                        uint8_t netmask[MTL_IP_ADDR_LEN]);
+
+int mt_socket_get_if_gateway(char* if_name, uint8_t gateway[MTL_IP_ADDR_LEN]);
 
 int mt_socket_get_if_mac(char* if_name, struct rte_ether_addr* ea);